% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{items_in_a_not_b}
\alias{items_in_a_not_b}
\title{https://stackoverflow.com/questions/10298662/find-elements-not-in-smaller-character-vector-list-but-in-big-list}
\usage{
items_in_a_not_b(a, b)
}
\description{
https://stackoverflow.com/questions/10298662/find-elements-not-in-smaller-character-vector-list-but-in-big-list
}
