% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files}
\alias{list_files}
\title{list all files in all subdirectories with a given extension}
\usage{
list_files(dir_path, ext = NULL, recursive = TRUE)
}
\description{
list all files in all subdirectories with a given extension
}
