% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_many_csv}
\alias{read_many_csv}
\title{Read all the csv files from a directory and append them into a single dataframe}
\usage{
read_many_csv(dir_path, sep = "\\t", row_names = 1)
}
\description{
Read all the csv files from a directory and append them into a single dataframe
}
