% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_index}
\alias{reset_index}
\title{See: https://stackoverflow.com/questions/36396911/r-move-index-column-to-first-column}
\usage{
reset_index(df, index_name = "index")
}
\description{
See: https://stackoverflow.com/questions/36396911/r-move-index-column-to-first-column
}
