% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_many_csv}
\alias{join_many_csv}
\title{Read all the csv files from a directory and left join them into a single dataframe
See: https://stackoverflow.com/questions/5319839/read-multiple-csv-files-into-separate-data-frames
by='id' or by=c('gene_id', 'gene_name', 'chromosome')}
\usage{
join_many_csv(dir_path, by, sep = "\\t")
}
\description{
Read all the csv files from a directory and left join them into a single dataframe
See: https://stackoverflow.com/questions/5319839/read-multiple-csv-files-into-separate-data-frames
by='id' or by=c('gene_id', 'gene_name', 'chromosome')
}
